Screen Adjustment Menu
Setting up Screen Adjustment Menu. Still needs functionality but i have the menu and the key bindings setup. It currently moves the screen but has no memory once returning, does not actually affect the screen outside this menu, and when it does, it needs to override the operation system aspect ratio and allow the screen to get bigger then it was or smaller then it was to fit tv screens (like CRT's) that dont perfectly fit windows aspect ratio choices.